Delete needs to be checked

+++ server/routes/user.js
@@ -96,7 +96,6 @@ exports.modifyUser = (req, res) => {
   //if (req.session.permission === 2) {
     if (req.method === 'POST') {
       const [permission] = [req.body.permission];
-      console.log(permission);
       const modifiedUser = [
         req.body.first_name,
         req.body.last_name,
@@ -130,14 +129,18 @@ exports.modifyUser = (req, res) => {
 exports.deleteUser = (req, res) => {
   const userId = req.params.id;
   if (req.method === 'POST') {
-    connection.query('DELETE FROM t_users WHERE ?', userId, (error) => {
-      if (error) {
-        const message = ('An error occured during the deletion : ', error);
-        res.render('delete', { message });
-      } else {
-        res.redirect('/user-management');
-      }
-    });
+    if (req.body.delete === 'yes') {
+      connection.query('DELETE FROM t_users WHERE ?', userId, (error) => {
+        if (error) {
+          const message = ('An error occured during the deletion : ', error);
+          res.render('delete', { message });
+        } else {
+          res.redirect('/user-management');
+        }
+      });
+    } else if (req.body.delete === 'no') {
+      res.redirect('/user-management');
+    }
   } else {
     connection.query('SELECT * FROM t_users WHERE id_user = ?', userId, (error, results) => {
       if (error) {
